Develop

+++ src/pages/CadCustomer/index.js
@@ -6,6 +6,7 @@ import HtmlTable from '../../components/HtmlTable'
 import HtmlModal from '../../components/HtmlModal'
 import ToastDisplay from '../../components/ToastDisplay'
 import PDFPrintTable from '../../functions/PdfPrintTable'
+import { confirmAlert } from 'react-confirm-alert'
 
 import {FiSearch} from 'react-icons/fi'
 import api from '../../services/api'
@@ -38,14 +39,29 @@ function CadCustomer(){
 	}, [setCustomers,authToken, q])
 
 	function deleteItem(id){
-		api.delete(`/customer/${id}`, {
-			headers: {
-				auth: authToken
-			}
-		}).then(()=>{
-			setCustomers(customers.filter(c=>parseInt(c.id)!==parseInt(id)))
-			setAlert(`Cliente ${id} deletado com sucesso!`)
-			setAlertChanger(alertChanger+1)
+		confirmAlert({
+			title: "Deletar Cliente",
+			message: `Deseja deletar Cliente ${id}?`,
+			buttons: [
+				{
+					label: 'Confirmar',
+					onClick: () => {
+						api.delete(`/customer/${id}`, {
+							headers: {
+								auth: authToken
+							}
+						}).then(()=>{
+							setCustomers(customers.filter(c=>parseInt(c.id)!==parseInt(id)))
+							setAlert(`Cliente ${id} deletado com sucesso!`)
+							setAlertChanger(alertChanger+1)
+						})
+					}
+				},
+				{
+					label: 'Cancelar',
+					onClick: () => {}
+				}
+			]		
 		})
 	}
 
@@ -60,7 +76,7 @@ function CadCustomer(){
 	}
 
 	function editItem(){
-		const selectedCustomer = customers.find(country=>country.selected)
+		const selectedCustomer = customers.find(customer=>customer.selected)
 
 		setEditedCustomer(selectedCustomer)
 
@@ -75,7 +91,7 @@ function CadCustomer(){
 	}
 
 	function newItem(){
-		setEditedCustomer(null)
+		setEditedCustomer(undefined)
 		
 		openModal()
 
@@ -99,7 +115,7 @@ function CadCustomer(){
 			setCustomers([...customers, customer])
 		}
 
-
+		setEditedCustomer(undefined)
 		closeModal()
 	}
 
@@ -129,9 +145,9 @@ function CadCustomer(){
 									{header:"Data de Nasc.", dataKey: "masked_birth_date"}
 								],
 								body: (customers.map(c=>{
-									c.masked_cpf = cpfMask(c.cpf)
-									c.masked_telephone = phoneMask(c.telephone)
-									c.masked_cep = cepMask(c.cep)
+									c.masked_cpf = cpfMask(c.ssa_vat_id||'')
+									c.masked_telephone = phoneMask(c.telephone||'')
+									c.masked_cep = cepMask(c.cep||'')
 									c.masked_birth_date = dateMask(c.birth_date)
 									return c
 								}))
@@ -149,10 +165,10 @@ function CadCustomer(){
 					{title:"Endereço", dataKey: "address"},
 					{title:"Data de Nasc.", dataKey: "masked_birth_date"},
 					]} tableData={customers.map(c=>{
-						c.masked_cpf = cpfMask(c.cpf)
-						c.masked_telephone = phoneMask(c.telephone)
-						c.masked_cep = cepMask(c.cep)
-						c.masked_birth_date = dateMask(c.birth_date)
+						c.masked_cpf = cpfMask(c.ssa_vat_id||'')
+						c.masked_telephone = phoneMask(c.telephone||'')
+						c.masked_cep = cepMask(c.cep||'')
+						c.masked_birth_date = dateMask(c.birth_date||'')
 						return c
 					})} selection={true} selectionCallback={items=>{
 						setCustomers(items)
